refactor: move PriceInfo to components
