+++ src/components/PriceInfo/index.tsx
