+++ src/components/RedeemUI/RedeemRequestStatusUI/RetriedRedeemRequest/index.tsx
@@ -7,7 +7,7 @@ import { FaExclamationCircle } from 'react-icons/fa';
 import { newMonetaryAmount } from '@interlay/interbtc-api';
 
 import RequestWrapper from 'pages/Bridge/RequestWrapper';
-import PriceInfo from 'pages/Bridge/PriceInfo';
+import PriceInfo from 'components/PriceInfo';
 import ExternalLink from 'components/ExternalLink';
 import PrimaryColorSpan from 'components/PrimaryColorSpan';
 import Hr2 from 'components/hrs/Hr2';
